+++ src/common.js
@@ -1,7 +1,5 @@
 const path = require('path')
-const homedir = /0.1/.test(process.version)
-  ? require('os-homedir')() // Node 0.12
-  : require('os').homedir() // Node 4+
+const homedir = require('os').homedir()
 
 const hotelDir = path.join(homedir, '.hotel')